Refactor+cleanup: moved all TestingBinary logic into utils.rs
