+++ src/validator.rs
@@ -1,8 +1,7 @@
 //! Module for the structs that represent and manage the validator/full-node processes i.e. Zebrad.
 use std::{
-    env,
     net::{IpAddr, Ipv4Addr, SocketAddr},
-    path::{Path, PathBuf},
+    path::PathBuf,
     process::Child,
 };
 
@@ -22,39 +21,12 @@ use crate::{
     error::LaunchError,
     launch, logs,
     network::{self, Network},
+    utils::{get_testing_bin_path, TestingBinary},
     Process,
 };
 
 /// Zebrad default miner address. Regtest/Testnet transparent address for [Abandon Abandon .. Art] seed (entropy all zeros)
 pub const ZEBRAD_DEFAULT_MINER: &str = "tmBsTi2xWTjUdEXnuTceL7fecEQKeWaPDJd";
-
-/// Testing binaries: these are managed and fetched automagically by fetcher.rs
-pub enum TestingBinary {
-    Lightwalletd,
-    Zainod,
-    ZcashCli,
-    Zcashd,
-    Zebrad,
-    ZingoCli,
-}
-
-/// Gets the right binary path for a certain Testing Binary
-pub fn get_testing_bin_path(binary: TestingBinary) -> PathBuf {
-    let crate_dir: String =
-        env::var("CARGO_MANIFEST_DIR").expect("cargo manifest path to be found");
-    let bins_dir = Path::new(&crate_dir).join("fetched_resources/test_binaries");
-
-    let name = match binary {
-        TestingBinary::Lightwalletd => "lightwalletd",
-        TestingBinary::Zainod => "zainod",
-        TestingBinary::ZcashCli => "zcash-cli",
-        TestingBinary::Zcashd => "zcashd",
-        TestingBinary::Zebrad => "zebrad",
-        TestingBinary::ZingoCli => "zingo-cli",
-    };
-
-    bins_dir.join(name).join("binary")
-}
 
 /// Zcashd configuration
 ///
